% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compteR.R
\docType{package}
\name{compteR}
\alias{compteR}
\alias{compteR-package}
\title{compteR: A package that helps groups to know who owes what to whom}
\description{
As the saying goes, "a debt paid is a friend kept". But for group spendings,
  determining who owes what to whom is a pain. \pkg{compteR} does it for you.
}
\details{
The dispatch function is likely the only function you will need to use. It produces a
  3-column table containing (i) the names of debtors and (ii) how much money
  they have to pay to reimburse (iii) their creditors.
}

