% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoPays.R
\name{whoPays}
\alias{whoPays}
\title{Splits a vector in 2 shorter ones depending on the amount a person should be reimbursed of.}
\usage{
whoPays(pos, neg)
}
\arguments{
\item{pos}{Numeric vector containing the amount of money some people in the
group have spend in excess and should be reimbursed of. Idealy, the vector
should have a name for each value.}

\item{neg}{Numeric vector containing the amount of money some people in the
group should reimburse to balance the spendings. Idealy, the vector should
have a name for each value.}
}
\value{
A list of 2 numerical vectors. \code{Remb} contains the values that a set of people need to spend to reimburse the first person in the \code{pos} vector. \code{Reste} contains the value that each debtor has still to spend to reimburse the remaining people in the \code{pos} vector.
}
\description{
This is an internal function. It should not be called directly by the user.
}
\examples{
pos <- c(125, 346)
names(pos) <- c("John", "Gail")
neg <- c(28, 312, 41, 90)
names(neg) <- c("Mike", "Lenny", "Christine", "Sally")
whoPays(pos,neg)
}

